Refactor Recife-PE spiders to allow multiple spiders
We have two different spiders so we can gather data from different
periods of time. Each spider is able to collect data from a specific
period of time.
